// src/API.ts
export type Favorites = {
  __typename: "Favorites",
  id: string,
  user?: string | null,
  jobID: Job,
  createdAt: string,
  updatedAt: string,
};

export type Job = {
  __typename: "Job",
  id: string,
  title?: string | null,
  description?: string | null,
  dateTimestamp?: number | null,
  dateDate?: string | null,
  location?: string | null,
  canBeDoneRemotely?: boolean | null,
  minimumPrice?: number | null,
  qualifications?: Array< string | null > | null,
  difficulty?: DifficultyEnums | null,
  status?: Status | null,
  urgent?: boolean | null,
  author?: User | null,
  Images?: ModelImagesConnection | null,
  needPreviousExperience?: boolean | null,
  categories?: Array< string | null > | null,
  points?: number | null,
  jobDoneBy?: JobDoneBy | null,
  userApply?: ModelUserApplyConnection | null,
  completedAt?: string | null,
  createdAt: string,
  updatedAt: string,
  jobAuthorId?: string | null,
  jobJobDoneById?: string | null,
};

export enum DifficultyEnums {
  EASY = "EASY",
  MEDIUM = "MEDIUM",
  HARD = "HARD",
  FAST = "FAST",
}


export enum Status {
  OPEN = "OPEN",
  INPROGRESS = "INPROGRESS",
  COMPLETED = "COMPLETED",
  ASSIGNED = "ASSIGNED",
  FINISHED = "FINISHED",
  CANCELED = "CANCELED",
  PAID = "PAID",
}


export type User = {
  __typename: "User",
  id: string,
  fullname?: string | null,
  reviewsLength?: number | null,
  totalJobsDone?: number | null,
  email?: string | null,
  phoneNumber?: string | null,
  avatar?: string | null,
  notificationsActived?: boolean | null,
  website?: string | null,
  firstTimeLoggin?: boolean | null,
  categories?: Array< string | null > | null,
  Locations?: ModelLocationConnection | null,
  Portifolios?: ModelPortifolioConnection | null,
  hasDeficiency?: boolean | null,
  Favorites?: ModelFavoritesConnection | null,
  ChatRooms?: ModelUserChatRoomConnection | null,
  Messages?: ModelMessageConnection | null,
  points?: number | null,
  profission?: string | null,
  backgroundImage?: string | null,
  Notifications?: ModelNotificationConnection | null,
  selectedPortifolio?: Portifolio | null,
  status?: UserStatus | null,
  jobsHistoric?: ModelJobHistoricModelConnection | null,
  ActivityHistoric?: ModelActivityHistoricConnection | null,
  about?: string | null,
  cv?: string | null,
  createdAt: string,
  updatedAt: string,
  userSelectedPortifolioId?: string | null,
};

export type ModelLocationConnection = {
  __typename: "ModelLocationConnection",
  items:  Array<Location | null >,
  nextToken?: string | null,
};

export type Location = {
  __typename: "Location",
  id: string,
  city?: string | null,
  state?: string | null,
  lat?: string | null,
  long?: string | null,
  country?: string | null,
  street?: string | null,
  userID?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelPortifolioConnection = {
  __typename: "ModelPortifolioConnection",
  items:  Array<Portifolio | null >,
  nextToken?: string | null,
};

export type Portifolio = {
  __typename: "Portifolio",
  id: string,
  title?: string | null,
  filesKey?: Array< string | null > | null,
  backgroundImageKey?: string | null,
  content?: string | null,
  userID: string,
  startedDate?: string | null,
  endDate?: string | null,
  techologiesUsed?: Array< string | null > | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelFavoritesConnection = {
  __typename: "ModelFavoritesConnection",
  items:  Array<Favorites | null >,
  nextToken?: string | null,
};

export type ModelUserChatRoomConnection = {
  __typename: "ModelUserChatRoomConnection",
  items:  Array<UserChatRoom | null >,
  nextToken?: string | null,
};
// ...
export type ModelMessageConnection = {
  __typename: "ModelMessageConnection",
  items:  Array<Message | null >,
  nextToken?: string | null,
};
// ...
export type ModelNotificationConnection = {
  __typename: "ModelNotificationConnection",
  items:  Array<Notification | null >,
  nextToken?: string | null,
};

export type Notification = {
  __typename: "Notification",
  id: string,
  text: string,
  userID: string,
  jobID: string,
  read?: boolean | null,
  senderID: string,
  notificationType?: NotificationType | null,
  createdAt: string,
  updatedAt: string,
};

export enum NotificationType {
  APPLIED = "APPLIED",
  COMPLETED = "COMPLETED",
  PAID = "PAID",
}


export enum UserStatus {
  REPORTED = "REPORTED",
  DELETED = "DELETED",
  FREQUENT = "FREQUENT",
  RARELY = "RARELY",
  NORMAL = "NORMAL",
}


export type ModelJobHistoricModelConnection = {
  __typename: "ModelJobHistoricModelConnection",
  items:  Array<JobHistoricModel | null >,
  nextToken?: string | null,
};

export type JobHistoricModel = {
  __typename: "JobHistoricModel",
  id: string,
  userID?: string | null,
  startDate?: string | null,
  endDate?: string | null,
  company?: string | null,
  jobTitle?: string | null,
  jobDescription?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type ModelActivityHistoricConnection = {
  __typename: "ModelActivityHistoricConnection",
  items:  Array<ActivityHistoric | null >,
  nextToken?: string | null,
};

export type ActivityHistoric = {
  __typename: "ActivityHistoric",
  id: string,
  userID: string,
  type?: ActivityType | null,
  typeID?: string | null,
  typeText?: string | null,
  createdAt: string,
  updatedAt: string,
};

export enum ActivityType {
  MESSAGE = "MESSAGE",
  JOB = "JOB",
  USER = "USER",
  PORTIFOLIO = "PORTIFOLIO",
}


export type ModelImagesConnection = {
  __typename: "ModelImagesConnection",
  items:  Array<Images | null >,
  nextToken?: string | null,
};

export type Images = {
  __typename: "Images",
  id: string,
  key?: string | null,
  url?: string | null,
  jobID?: string | null,
  createdAt: string,
  updatedAt: string,
};

export type JobDoneBy = {
  __typename: "JobDoneBy",
  id: string,
  jobID?: string | null,
  userID?: string | null,
  user?: User | null,
  status?: Status | null,
  paid?: boolean | null,
  jobPaidValue?: string | null,
  jobPaidDate?: string | null,
  createdAt: string,
  updatedAt: string,
  jobDoneByUserId?: string | null,
};

export type ModelUserApplyConnection = {
  __typename: "ModelUserApplyConnection",
  items:  Array<UserApply | null >,
  nextToken?: string | null,
};

export type UserApply = {
  __typename: "UserApply",
  id: string,
  userID?: string | null,
  jobID?: string | null,
  selectedPortifolio?: Portifolio | null,
  files?: Array< string | null > | null,
  text?: string | null,
  price?: string | null,
  userName?: string | null,
  userAvatar?: string | null,
  userStatus?: UserStatus | null,
  createdAt: string,
  updatedAt: string,
  userApplySelectedPortifolioId?: string | null,
};